fix(prefer-const): remove unnecessary `FixTracker`
Code has been moved into `prefer-const-helpers` folder, so the rule file
is simpler.
